t-36: review fix
t-36: fix testCaseDocumentController logs

t-36: fix download

Fix test "project/delete"

Fix migrations

t-36: review fix

+++ grails-app/controllers/testingforest/ProjectController.groovy
@@ -101,9 +101,13 @@ class ProjectController {
     }
 
     def delete(Long projectId) {
-        log.info("Deleting ${projectService.get(projectId).projectName}")
+        Project project = Project.get(projectId)
+        if(project) {
+            log.info("Deleting ${project.projectName}")
 
-        projectService.delete(projectId)
-        redirect uri: "/project/index"
+            projectService.delete(projectId)
+
+            redirect uri: "/project/index"
+        }
     }
 }
